tiff-ai-analyzer/requirements.txt
Pillow==10.0.0
openai==1.3.0
python-xmp-toolkit==2.0.1
exifread==3.0.0
python-dotenv==1.0.0
